api/requirements.txt
fastapi
ruamel.yaml
pydantic
uvicorn
docker